// worker/src/replicache-protocol-schemas.ts
// Generated by ts-to-zod
import { z } from "zod";

const mutationSchema = z.object({
  id: z.number(),
  name: z.string(),
  args: z.any().optional(),
});

export const pushRequestBodySchema = z.object({
  clientID: z.string(),
  mutations: z.array(mutationSchema),
  pushVersion: z.number(),
  schemaVersion: z.string(),
});